modularize bert tokenizer
