+++ bert/examples/bert.rs
@@ -14,12 +14,13 @@
 
 //! Run as `cargo run --example bert
 
-use xayn_ai_bert::{Config, FirstPooler};
+use xayn_ai_bert::{tokenizer::bert::Tokenizer, Config, FirstPooler};
 use xayn_ai_test_utils::asset::smbert;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let pipeline = Config::new(smbert()?)?
         .with_token_size(64)?
+        .with_tokenizer::<Tokenizer>()
         .with_pooler::<FirstPooler>()
         .build()?;
     let embedding = pipeline.run("This is a sequence.")?;
